pyproject.toml
[tool.poetry]
name = "dinner-bell"
version = "0.1.0"
description = "An operator to notify you when a deployment is ready"
authors = ["Dann Bohn <djb44@psu.edu>"]

[tool.poetry.dependencies]
python = "^3.7"
kopf = "^0.27"
kubernetes = "^11.0.0"
slackclient = "^2.8.0"
requests = "^2.24.0"
black = "^19.10b0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
